# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Draft App — card library and MTG-style draft
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drafts:
    get:
      operationId: listDrafts
      tags: [drafts]
      summary: List all draft sessions
      responses:
        "200":
          description: List of drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draft"
    post:
      operationId: createDraft
      tags: [drafts]
      summary: Create a new draft session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftInput"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}:
    get:
      operationId: getDraft
      tags: [drafts]
      summary: Get draft details with seats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft with seats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteDraft
      tags: [drafts]
      summary: Delete a draft and all associated data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Draft deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}/join:
    post:
      operationId: joinDraft
      tags: [drafts]
      summary: Join a draft as a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinDraftInput"
      responses:
        "201":
          description: Joined — returns seat info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seat"
        "400":
          description: Invalid or draft already started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Draft not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}/start:
    post:
      operationId: startDraft
      tags: [drafts]
      summary: Start the draft — distribute cards into packs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftDetail"
        "400":
          description: Cannot start (wrong state, too few cards, no players)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Draft not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}/seat/{seatId}:
    get:
      operationId: getSeatState
      tags: [drafts]
      summary: Get current pack for a seat (what the player sees)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: seatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Current seat state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatState"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}/pick:
    post:
      operationId: makePick
      tags: [drafts]
      summary: Pick a card from the current pack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickInput"
      responses:
        "200":
          description: Pick recorded, returns updated seat state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatState"
        "400":
          description: Invalid pick
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}/pool/{seatId}:
    get:
      operationId: getPool
      tags: [drafts]
      summary: Get the full pool of cards picked by a seat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: seatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Picked card pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}/all-pools:
    get:
      operationId: getAllPools
      tags: [drafts]
      summary: Get every seat's picked card pool for a completed draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: All pools
          content:
            application/json:
              schema:
                type: object
                required: [seatPools]
                properties:
                  seatPools:
                    type: array
                    items:
                      $ref: "#/components/schemas/Pool"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /drafts/{id}/summary:
    get:
      operationId: getDraftSummary
      tags: [drafts]
      summary: Get summary of draft progress across all seats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Card:
      type: object
      required: [id, name, imageObjectPath, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        imageObjectPath:
          type: string
          description: Path to the stored image (e.g. /objects/uploads/uuid)
        createdAt:
          type: string
          format: date-time
...
    Draft:
      type: object
      required: [id, name, status, numPacks, cardsPerPack, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [waiting, active, completed]
        numPacks:
          type: integer
        cardsPerPack:
          type: integer
        createdAt:
          type: string
          format: date-time
    DraftInput:
      type: object
      required: [name, numPacks, cardsPerPack]
      properties:
        name:
          type: string
          minLength: 1
        numPacks:
          type: integer
          minimum: 1
        cardsPerPack:
          type: integer
          minimum: 1
    Seat:
      type: object
      required: [id, draftId, playerName, seatPosition, joinedAt]
      properties:
        id:
          type: integer
        draftId:
          type: integer
        playerName:
          type: string
        seatPosition:
          type: integer
        joinedAt:
          type: string
          format: date-time
    JoinDraftInput:
      type: object
      required: [playerName]
      properties:
        playerName:
          type: string
          minLength: 1
    DraftDetail:
      type: object
      required: [id, name, status, numPacks, cardsPerPack, createdAt, seats]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [waiting, active, completed]
        numPacks:
          type: integer
        cardsPerPack:
          type: integer
        createdAt:
          type: string
          format: date-time
        seats:
          type: array
          items:
            $ref: "#/components/schemas/Seat"
    SeatState:
      type: object
      required: [seat, currentPack, packNumber, picksDone, totalPicks, draftStatus]
      properties:
        seat:
          $ref: "#/components/schemas/Seat"
        currentPack:
          type: array
          items:
            $ref: "#/components/schemas/Card"
          description: Cards remaining in the current pack (empty if waiting for pack to arrive or draft done)
        currentPackId:
          type: integer
          description: The ID of the pack currently shown (present when currentPack is non-empty)
        packNumber:
          type: integer
          description: Which pack number is currently being drafted (1-based)
        picksDone:
          type: integer
          description: Total picks this seat has made
        totalPicks:
          type: integer
          description: Total picks this seat needs to make (numPacks * cardsPerPack)
        draftStatus:
          type: string
          enum: [waiting, active, completed]
        waitingForPack:
          type: boolean
          description: True when this seat has picked but is waiting for the passed pack to arrive
        waitingForOthers:
          type: boolean
          description: True when this seat has already picked this round and is waiting for the other players to pick before packs rotate
    PickInput:
      type: object
      required: [seatId, cardId, packId]
      properties:
        seatId:
          type: integer
        cardId:
          type: integer
        packId:
          type: integer
          description: The specific pack ID being picked from (must match the pack currently assigned to the seat)
    Pool:
      type: object
      required: [seat, cards]
      properties:
        seat:
          $ref: "#/components/schemas/Seat"
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
    SeatSummary:
      type: object
      required: [seat, picksDone, waitingForPack]
      properties:
        seat:
          $ref: "#/components/schemas/Seat"
        picksDone:
          type: integer
        waitingForPack:
          type: boolean
    DraftSummary:
      type: object
      required: [draft, totalPicksMade, totalPicksNeeded, seatSummaries]
      properties:
        draft:
          $ref: "#/components/schemas/Draft"
        totalPicksMade:
          type: integer
        totalPicksNeeded:
          type: integer
        seatSummaries:
          type: array
          items:
            $ref: "#/components/schemas/SeatSummary"
...